% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hola.R
\name{hola}
\alias{hola}
\title{Greetings in Spanish}
\usage{
hola(nombre)
}
\arguments{
\item{nombre}{The name of the person we are saying hi to in Spanish.}
}
\value{
A \code{character(1)} with the greetings in Spanish
}
\description{
This function says hi to people
}
